split clock system into public and admin modules
