+++ packages/react-frontend/src/app/(public)/layout.tsx
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function PublicLayout({ children }: { children: React.ReactNode }) {
+  return <div className="min-h-screen bg-white">{children}</div>
+}
