+++ packages/react-frontend/src/app/(public)/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import PublicClockSystem from "../../public-clock-system"
+
+export default function Page() {
+  return <PublicClockSystem />
+}
