+++ packages/react-frontend/src/app/admin/layout.tsx
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function AdminLayout({ children }: { children: React.ReactNode }) {
+  return <div className="min-h-screen bg-slate-50">{children}</div>
+}
