+++ packages/react-frontend/src/app/admin/page.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import AdminClockSystem from "../../admin-clock-system"
+
+export default function AdminPage() {
+  return <AdminClockSystem />
+}
