--- packages/react-frontend/src/app/page.tsx
@@ -1,7 +1,0 @@
-"use client"
-
-import Component from "../clock-system"
-
-export default function Page() {
-  return <Component />
-}
