+++ packages/react-frontend/src/data/initialData.ts
@@ -1,0 +1,171 @@
+export const availableLocations = [
+  "Help Desk - Main",
+  "Help Desk - Library",
+  "Lab Support - A",
+  "Lab Support - B",
+  "Remote Support",
+  "Hardware Repair",
+  "Software Support",
+  "Network Operations",
+]
+
+export const initialTerms = [
+  {
+    id: "1",
+    name: "Fall 2025",
+    startDate: "2025-08-25",
+    endDate: "2025-12-15",
+    isActive: true,
+  },
+  {
+    id: "2",
+    name: "Summer 2025",
+    startDate: "2025-05-15",
+    endDate: "2025-08-15",
+    isActive: false,
+  },
+  {
+    id: "3",
+    name: "Spring 2025",
+    startDate: "2025-01-15",
+    endDate: "2025-05-10",
+    isActive: false,
+  },
+  {
+    id: "4",
+    name: "Thanksgiving Break 2025",
+    startDate: "2025-11-25",
+    endDate: "2025-11-29",
+    isActive: false,
+  },
+]
+
+export const initialStaffData = [
+  {
+    id: 1,
+    name: "Alex Chen",
+    cardId: "CARD001",
+    role: "Student Lead",
+    currentStatus: "present",
+    assignedLocation: "Help Desk - Main", // Assigned location
+    weeklySchedule: {
+      monday: ["8:30-12:00", "1:00-5:00"],
+      tuesday: ["9:00-1:00"],
+      wednesday: ["8:30-12:00", "1:00-5:00"],
+      thursday: ["9:00-1:00"],
+      friday: ["8:30-12:00"],
+      saturday: [],
+      sunday: [],
+    },
+    todayActual: "08:25 AM",
+    clockEntries: [
+      { timestamp: "2025-01-20T08:25:00", type: "in" },
+      { timestamp: "2025-01-20T12:30:00", type: "out" },
+      { timestamp: "2025-01-20T13:15:00", type: "in" },
+      { timestamp: "2025-01-20T17:05:00", type: "out" },
+    ],
+  },
+  {
+    id: 2,
+    name: "Sarah Johnson",
+    cardId: "CARD002",
+    role: "Assistant",
+    currentStatus: "present",
+    assignedLocation: "Lab Support - A", // Assigned location
+    weeklySchedule: {
+      monday: ["9:15-5:00"],
+      tuesday: ["9:15-5:00"],
+      wednesday: ["9:15-5:00"],
+      thursday: ["9:15-5:00"],
+      friday: ["9:15-1:00"],
+      saturday: [],
+      sunday: [],
+    },
+    todayActual: "09:10 AM",
+    clockEntries: [{ timestamp: "2025-01-20T09:10:00", type: "in" }],
+  },
+  {
+    id: 3,
+    name: "Mike Rodriguez",
+    cardId: "CARD003",
+    role: "Student Lead",
+    currentStatus: "present",
+    assignedLocation: undefined, // Pending assignment
+    weeklySchedule: {
+      monday: ["8:45-12:00", "1:00-4:00"],
+      tuesday: ["10:00-2:00"],
+      wednesday: ["8:45-12:00", "1:00-4:00"],
+      thursday: ["10:00-2:00"],
+      friday: ["8:45-12:00"],
+      saturday: [],
+      sunday: [],
+    },
+    todayActual: "08:45 AM",
+    clockEntries: [
+      { timestamp: "2025-01-20T08:45:00", type: "in" },
+      { timestamp: "2025-01-20T12:15:00", type: "out" },
+      { timestamp: "2025-01-20T13:10:00", type: "in" },
+    ],
+  },
+  {
+    id: 4,
+    name: "Emma Wilson",
+    cardId: "CARD004",
+    role: "Assistant",
+    currentStatus: "expected",
+    assignedLocation: undefined, // Not clocked in yet
+    weeklySchedule: {
+      monday: ["9:00-5:00"],
+      tuesday: ["9:00-5:00"],
+      wednesday: ["9:00-1:00"],
+      thursday: ["9:00-5:00"],
+      friday: ["9:00-5:00"],
+      saturday: [],
+      sunday: [],
+    },
+    todayActual: null,
+    clockEntries: [],
+  },
+  {
+    id: 5,
+    name: "David Park",
+    cardId: "CARD005",
+    role: "Assistant",
+    currentStatus: "expected",
+    assignedLocation: undefined, // Not clocked in yet
+    weeklySchedule: {
+      monday: [],
+      tuesday: ["10:30-2:30"],
+      wednesday: ["10:30-2:30"],
+      thursday: ["10:30-2:30"],
+      friday: [],
+      saturday: [],
+      sunday: [],
+    },
+    todayActual: null,
+    clockEntries: [],
+  },
+  {
+    id: 6,
+    name: "Lisa Zhang",
+    cardId: "CARD006",
+    role: "Student Lead",
+    currentStatus: "present",
+    assignedLocation: "Network Operations", // Assigned location
+    weeklySchedule: {
+      monday: ["8:00-12:00", "1:00-5:00"],
+      tuesday: ["8:00-12:00", "1:00-5:00"],
+      wednesday: ["8:00-12:00", "1:00-5:00"],
+      thursday: ["8:00-12:00", "1:00-5:00"],
+      friday: ["8:00-12:00"],
+      saturday: [],
+      sunday: [],
+    },
+    todayActual: "07:58 AM",
+    clockEntries: [
+      { timestamp: "2025-01-20T07:58:00", type: "in" },
+      { timestamp: "2025-01-20T12:05:00", type: "out" },
+    ],
+  },
+]
+
